Add support for deletion of wan replicated maps
Added finalizer to maps for which we start Wan replication. So, when
the map is deleted Wan reconciler has the time to do the clean up. I
also added owner reference to the maps along with a map watcher to
Wan reconciler manager. This is done to be able to trigger the Wan
reconciler when a replicated map is marked to be deleted.

Along with the support, did the following:
- reordered functions top down approach
- fixed a bug where the error was hidden in stopWanReplicationAllMaps function
- removed duplicate entries in Hazelcast Map list
- Moved stopWanRepForRemovedResources from inside appliedBefore because partial applications can exist
- Fixed stopWanRepForRemovedResources function where we were empty map
- Changed map deletion mode in stopWanReplicationForAllMaps and stopWanRepForRemovedResources, this is done to make finalizer and map status deletion as close as possible.