[CN-1290] Check for operator pod's status and restart count after test suites (#1165)
## Description

It helps us to detect which of the test cases leading operator pod to be
restarted. It will make the workflow failed if the operator pod's
restart count is not 0 after all test suited executed. Then, we can
simply check the events of the related tests in Grafana board to have an
insight.

## User Impact

<!--- Please describe any user facing impact of this change. This can be
positive or negative impact. -->